refactor(fix): code quality improvements from /improve review
- FIX framer: reuse compute_checksum() from parser.hpp instead of
  duplicating the byte-sum logic (DRY principle)
- FIX framer: fix loop variable type (int → size_t) for consistency
- FIX parser: remove unused #include <algorithm>
- Benchmarks: fix loop variable types (int → size_t) for C++23 style

+++ benchmarks/fix/bench_fix_parse.cpp
@@ -58,7 +58,7 @@ std::vector<uint8_t> build_market_data_snapshot() {
     b.set(tag::Symbol, "TSLA");
     // Repeating group: 5 MD entries
     b.set_int(268, 5);  // NoMDEntries
-    for (int i = 0; i < 5; ++i) {
+    for (size_t i = 0; i < 5; ++i) {
         b.set_int(269, i % 2);  // MDEntryType: 0=Bid, 1=Offer
         b.set_double(270, 200.00 + i * 0.25, 2);  // MDEntryPx
         b.set_int(271, (i + 1) * 100);  // MDEntrySize
